requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0